+++ apps/backend/src/libs/modules/aws/libs/enums/enums.ts
@@ -1,0 +1,1 @@
+export { AwsExceptionMessage } from "./aws-file.exeption-message.js";