+++ apps/backend/src/modules/auth/auth.controller.test.ts
@@ -33,6 +33,7 @@ describe("AuthController", () => {
 		},
 		groupId: 2,
 		id: 1,
+		isVisibleProfile: true,
 		lastName: "Doe",
 	};
 
@@ -50,6 +51,7 @@ describe("AuthController", () => {
 		} as GroupItemWithPermissionsDto,
 		groupId: 2,
 		id: 1,
+		isVisibleProfile: true,
 		lastName: "Smith",
 	};