+++ apps/frontend/src/libs/hooks/hooks.ts
@@ -1,5 +1,7 @@
 export { useAppDispatch } from "./use-app-dispatch/use-app-dispatch.hook.js";
+export { useAppEntryRedirect } from "./use-app-entry-redirect/use-app-entry-redirect.hook.js";
 export { useAppForm } from "./use-app-form/use-app-form.hook.js";
+export { useAppNavigate } from "./use-app-navigate/use-app-navigate.hook.js";
 export { useAppSelector } from "./use-app-selector/use-app-selector.hook.js";
 export { useMapClient } from "./use-map-client/use-map-client.hook.js";
 export { useTabNavigation } from "./use-tab-navigation/use-tab-navigation.hook.js";