+++ apps/frontend/src/libs/modules/store/store.module.ts
@@ -9,6 +9,7 @@ import { AppEnvironment } from "~/libs/enums/enums.js";
 import { type Config } from "~/libs/modules/config/config.js";
 import { storage } from "~/libs/modules/storage/storage.js";
 import { toastNotifier } from "~/libs/modules/toast-notifier/toast-notifier.js";
+import { reducer as appReducer } from "~/modules/app/app.js";
 import { authApi, reducer as authReducer } from "~/modules/auth/auth.js";
 import {
 	pointOfInterestApi,
@@ -16,7 +17,7 @@ import {
 	pointsOfInterestReducer,
 } from "~/modules/points-of-interest/points-of-interest.js";
 import { routeApi, reducer as routeReducer } from "~/modules/routes/routes.js";
-import { userApi, reducer as usersReducer } from "~/modules/users/users.js";
+import { userApi } from "~/modules/users/users.js";
 
 import { type ExtraArguments, type RootReducer } from "./libs/types/types.js";
 import { handleErrorMiddleware } from "./middlewares/middlewares.js";
@@ -52,11 +53,11 @@ class Store {
 				}).prepend(handleErrorMiddleware(this.extraArguments));
 			},
 			reducer: {
+				app: appReducer,
 				auth: authReducer,
 				pointOfInterestDetails: pointOfInterestDetailsReducer,
 				pointsOfInterest: pointsOfInterestReducer,
 				route: routeReducer,
-				users: usersReducer,
 			},
 		});
 	}