+++ apps/frontend/src/pages/route-description/route-details.tsx
@@ -2,16 +2,17 @@ import image1 from "~/assets/images/route-details/placeholder-image-1.png";
 import image2 from "~/assets/images/route-details/placeholder-image-2.png";
 import image3 from "~/assets/images/route-details/placeholder-image-3.png";
 import { Header, ImageGallery, Loader } from "~/libs/components/components.js";
-import { AppRoute } from "~/libs/enums/app-route.enum.js";
-import { DataStatus } from "~/libs/enums/data-status.enum.js";
+import { AppRoute, DataStatus } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
 	useAppSelector,
 	useEffect,
 	useParams,
 } from "~/libs/hooks/hooks.js";
-import { type RouteGetByIdResponseDto } from "~/libs/types/types.js";
-import { actions as routeActions } from "~/modules/routes/routes.js";
+import {
+	actions as routeActions,
+	type RouteGetByIdResponseDto,
+} from "~/modules/routes/routes.js";
 
 import { NotFound } from "../not-found/not-found.js";
 import styles from "./styles.module.css";
@@ -29,7 +30,7 @@ const RouteDetails = (): React.JSX.Element => {
 		void dispatch(routeActions.getRouteById(Number(id)));
 	}, [dispatch, id]);
 
-	const data = useAppSelector(({ route }) => route.route);
+	const route = useAppSelector(({ route }) => route.route);
 
 	const dataStatus = useAppSelector(({ route }) => route.dataStatus);
 
@@ -41,7 +42,7 @@ const RouteDetails = (): React.JSX.Element => {
 		return <Loader />;
 	}
 
-	const { description, name } = data as RouteGetByIdResponseDto;
+	const { description, name } = route as RouteGetByIdResponseDto;
 
 	return (
 		<>