+++ packages/shared/src/libs/enums/api-path.enum.ts
@@ -1,5 +1,6 @@
 const APIPath = {
 	AUTH: "/auth",
+	FILES: "/files",
 	NOTIFICATIONS: "/notifications",
 	POINTS_OF_INTEREST: "/points-of-interest",
 	POINTS_OF_INTEREST_$ID: "/points-of-interest/:id",