+++ packages/shared/src/modules/users/libs/types/authenticated-user-patch-response-dto.type.ts
@@ -6,6 +6,7 @@ type AuthenticatedUserPatchResponseDto = {
 	group: GroupItemWithPermissionsDto;
 	groupId: number;
 	id: number;
+	isVisibleProfile: boolean;
 	lastName: string;
 };